backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
httpx==0.27.2
python-multipart==0.0.12
aiosmtplib==3.0.2
transliterate==1.10.2
aiofiles==24.1.0
